# servers/swagger-codegen/api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple TODO
  description: A simple TODO application
  contact:
    name: David Herron
    url: https://davidherron.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
- url: /
paths:
  /todo:
    post:
      summary: Create a TODO
      operationId: createTodo
      requestBody:
        description: Parameters for creating a TODO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoPost'
      responses:
        "200":
          description: Successful
        "400":
          description: Invalid status value
      x-swagger-router-controller: Default
  /todos/all:
    get:
      summary: List all TODO items
      description: ""
      operationId: listTodos
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
                x-content-type: application/json
        "400":
          description: Invalid status value
      x-swagger-router-controller: Default
  /todos/{id}:
    get:
      summary: Find TODO by ID
      description: Returns a single TODO
      operationId: getTodoById
      parameters:
      - name: id
        in: path
        description: ID of Todo to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        "400":
          description: Invalid ID supplied
        "404":
          description: TODO not found
      x-swagger-router-controller: Default
    delete:
      summary: Deletes a TODO
      description: ""
      operationId: deleteTodo
      parameters:
      - name: id
        in: path
        description: TODO id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful deletion
        "400":
          description: Invalid ID supplied
        "404":
          description: TODO not found
      x-swagger-router-controller: Default
    patch:
      summary: Update an existing TODO
      operationId: patchTodoById
      parameters:
      - name: id
        in: path
        description: TODO id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoPatch'
      responses:
        "200":
          description: Success
        "400":
          description: Invalid ID supplied
        "404":
          description: TODO not found
      x-swagger-router-controller: Default
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        precedence:
          type: string
          enum:
          - low
          - medium
          - high
        complete:
          type: boolean
          default: false
      example:
        id: id
        title: title
        body: body
        complete: false
        precedence: low
    TodoPost:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        precedence:
          type: string
          enum:
          - low
          - medium
          - high
    TodoPatch:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        precedence:
          type: string
          enum:
          - low
          - medium
          - high